requirements.txt
streamlit
python-dotenv
jamaibase